backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
MLB-StatsAPI==1.6.2
python-multipart==0.0.6